#1087 - Add support to affordances for input and output media types.
Adds the ability to specify input media types and output media types in the AffordanceBuilder APIs. This is, in turn, used by SpringAffordanceBuilder and applying the `consumes` and `produces` clauses for Spring MVC/WebFlux's annotations. HAL-FORMS in turn can populate `contentType` using this information. Other media types are free to leverage this additional information as well.